// app/schemas/auth.ts
import { email } from "better-auth";
import z from "zod";
export const signUpSchema = z.object({
    name: z.string().min(3).max(30),
    email: z.string().email(),
    password: z.string().min(8).max(12)
})
export const LoginSchema = z.object({
    email: z.string().email(),
    password: z.string().min(8).max(12)
})